Separated the controller of posts and comments
The post controller is now in PostsController.java and the comment controller is in CommentsController.java. Also, posts and comments about which the user complained will be hidden for this user. A function has been created to check whether the user has complained about this post or comment, if so he will not be able to send the complaint again.